Refactorizing API (phase1)
