+++ src/api(old)/api.py
@@ -1,107 +1,3 @@
-"""
-API - MAIN
-"""
-# Version : 0.2.0
-# Current state : Dev
-# Author : Guillaume Pot
-# Contact : guillaumepot.pro@outlook.com
-api_version = "0.2.0"
-current_state = "Dev"
-
-
-"""
-LIB
-"""
-import os
-import jwt
-import uuid
-import pandas as pd
-from fastapi import FastAPI, Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from passlib.context import CryptContext
-from datetime import datetime, timedelta
-
-from sqlalchemy import create_engine
-
-
-
-"""
-VARS
-"""
-# Auth
-crypt_context_scheme = os.getenv("CRYPT_CONTEXT_SCHEME")
-pwd_context = CryptContext(schemes=[crypt_context_scheme], deprecated="auto")
-
-authorized_users = os.getenv("AUTHORIZED_USERS").split(',')
-
-access_token_expiration = int(os.getenv("ACCESS_TOKEN_EXPIRATION"))
-algorithm = os.getenv("ALGORITHM")
-jwt_secret_key = os.getenv("BANK_APP_API_TOKEN_SECRET_KEY")
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl=f"/api/{api_version}/login")
-
-
-# Database related variables
-host = os.getenv('POSTGRES_HOST')
-port = os.getenv('POSTGRES_PORT')
-user = os.getenv('POSTGRES_USER')
-password = os.getenv('POSTGRES_PASSWORD')
-database = os.getenv('POSTGRES_DB')
-
-engine = create_engine(f"postgresql://{user}:{password}@{host}:{port}/{database}")
-
-
-# Value Locks
-available_account_types = ("checking", "saving")
-available_transactions_types = ("debit", "credit", "transfert")
-
-
-"""
-API declaration
-"""
-app = FastAPI(
-    title="Personal bank app",
-    description=f"Personal bank app - API {api_version}",
-    version=f"{current_state}",
-    openapi_tags=[
-        {
-            'name': 'auth',
-            'description': 'Authentication'
-        },
-        {
-            'name': 'home',
-            'description': 'Home'
-        },
-        {
-            'name': 'account',
-            'description': 'Account'
-        },
-        {
-            'name': 'budget',
-            'description': 'Budget'
-        },
-        {
-            'name': 'transaction',
-            'description': 'Transaction'
-        }   
-    ]
-)
-
-###############################################################
-"""
-API Logger
-"""
-
-# LOGGER WIP
-
-
-
-######################################################################################
-"""
-FUNCTIONS
-"""
-
-
 
 ######################################################################################
 """
@@ -176,23 +72,7 @@ def log_user(credentials: OAuth2PasswordRequestForm = Depends()):
     return {"access_token": encoded_jwt}
 
 ######################################################################################
-"""
-ROUTES
-"""
 
-"""
-home routes
-"""
-# status route
-@app.get(f"/api/{api_version}/status", name="status", tags=['home'])
-async def get_status() -> dict:
-    """
-    Get the status of the API.
-
-    Returns:
-        dict: A dictionary containing the status, version, and current state of the API.
-    """
-    return {"status": "Working", "version": api_version, "current_state": current_state}
